requirements.txt
# Core ML libraries
torch>=2.0.0
transformers>=4.46.0
datasets>=2.14.0
accelerate>=0.24.0

# GRPO and RL training
trl>=0.12.0
peft>=0.7.0

# Quantization
bitsandbytes>=0.41.0

# Image processing
Pillow>=9.0.0
pydicom>=2.3.0  # For DICOM medical images

# Utilities
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Optional: for faster training
flash-attn>=2.0.0  # Optional, for faster attention